[SOIN] Déplace le type "Notification" dans le dossier 'ui'
... car le type est maintenant attaché au `storeNotification`, et pourra être utilisé par plusieurs composants
