+++ svelte/lib/ui/types.d.ts
@@ -7,3 +7,29 @@ export enum Referentiel {
 export type IdStatut = string;
 
 export type ReferentielStatut = Record<IdStatut, string>;
+
+export type TypeNotification = 'nouveaute' | 'tache';
+
+type NotificationBase = {
+  id: string;
+  titre: string;
+  statutLecture: 'lue' | 'nonLue';
+  lien: string;
+  type: TypeNotification;
+  doitNotifierLecture: boolean;
+  horodatage?: string;
+};
+
+export type NotificationNouveaute = NotificationBase & {
+  type: 'nouveaute';
+  image: string;
+  sousTitre: string;
+};
+
+export type NotificationTache = NotificationBase & {
+  type: 'tache';
+  titreCta: string;
+  entete: string;
+};
+
+export type Notification = NotificationNouveaute | NotificationTache;
